+++ R/app_ui.R
@@ -2,7 +2,8 @@
 #'
 #' @param request Internal parameter for `{shiny}`.
 #'     DO NOT REMOVE.
-#' @import shiny
+#' @rawNamespace import(shiny, except = c(dataTableOutput, renderDataTable))
+#' @importFrom utils head installed.packages packageVersion write.csv zip 
 #' @importFrom shinymanager secure_app
 #' 
 #' @noRd
@@ -10,7 +11,7 @@ app_ui <- function(request) {
   # Your application UI logic
   ui <- fluidPage(
     
-    theme = app_theme, # defined in data-raw/interanl-data.R
+    theme = app_theme(), # defined in data-raw/interanl-data.R
     
     # not needed any more. Automatically bundled below
     # includeCSS(path = file.path('www', 'css', 'main.css')),
@@ -76,7 +77,7 @@ app_ui <- function(request) {
     wellPanel(
       id = "footer",
       "Checkout the app's code!",
-      tags$a(href = "https://github.com/pharmaR/risk_assessment",
+      tags$a(href = "https://github.com/pharmaR/riskassessment",
              icon("github-alt"), target = "_blank")
     )
     
@@ -94,11 +95,10 @@ app_ui <- function(request) {
 #' This function is internally used to add external
 #' resources inside the Shiny application.
 #'
-#' @import shiny
+#' 
 #' @importFrom golem add_resource_path activate_js favicon bundle_resources
 #' @importFrom rintrojs introjsUI
 #' @importFrom shinyjs useShinyjs
-#' @importFrom waiter use_waitress
 #' @importFrom golem add_resource_path activate_js favicon bundle_resources
 #' @noRd
 golem_add_external_resources <- function() {
@@ -116,6 +116,5 @@ golem_add_external_resources <- function() {
     # Add here other external resources
     rintrojs::introjsUI(),
     shinyjs::useShinyjs(),
-    waiter::use_waitress(),
   )
 }